Fix needing -fpermissive
FossilOrigin-Name: bd82b5a5b17fdeacf6b3133104b958a1597e5e16379280cf0bcfef7e7fef1468

+++ src/clients2c.cxx
@@ -273,7 +273,8 @@ localservertoclient(
 			setspawn(i, true);
 			if (i >= (uint)ents.length())
 				break;
-			vec v = {ents[i].x, ents[i].y, ents[i].z};
+			vec v = {(float)ents[i].x, (float)ents[i].y,
+			    (float)ents[i].z};
 			playsound(S_ITEMSPAWN, &v);
 			break;
 		};
