+++ src/world.cxx
@@ -256,7 +256,7 @@ closestent() // used for delent and edit mode ent display
 		entity &e = ents[i];
 		if (e.type == NOTUSED)
 			continue;
-		vec v = {e.x, e.y, e.z};
+		vec v = {(float)e.x, (float)e.y, (float)e.z};
 		vdist(dist, t, player1->o, v);
 		if (dist < bdist) {
 			best = i;
@@ -316,7 +316,8 @@ entity *
 newentity(int x, int y, int z, char *what, int v1, int v2, int v3, int v4)
 {
 	int type = findtype(what);
-	persistent_entity e = {x, y, z, v1, type, v2, v3, v4};
+	persistent_entity e = {(short)x, (short)y, (short)z, (short)v1,
+	    (uchar)type, (uchar)v2, (uchar)v3, (uchar)v4};
 	switch (type) {
 	case LIGHT:
 		if (v1 > 32)
